resolve and remove TODOs from extra tests
import() from the exporter was being cleaned before it got called.

We should strongly consider special-casing import and unimport, since as they
are always called as class methods from outside the class, it never makes
sense to clean them, and it is difficult to install them into the package
(Sub::Name adds an extra stack frame, which interferes with figuring out where
to import the symbols to).